v490.7-dev 2 pre-processing updates.

+++ scripts/ANALYSIS.evndisp.sh
@@ -134,7 +134,7 @@ echo
 # sleep required for large data sets to avoid overload
 # of database and many jobs running in parallel
 SLEEPABIT="1s"
-if [ "$NRUNS" -gt "50" ] ; then
+if [ "$NRUNS" -gt "50" ] && [[ ${DBTEXTDIR} == "0" ]]; then
    SLEEPABIT="30s"
    echo "Long list of runs (${NRUNS}), will sleep after each run for ${SLEEPABIT}"
 fi
@@ -146,13 +146,17 @@ do
     mkdir -p ${ODIR}/Calibration/${T}
 done
 if [[ -e "${VERITAS_EVNDISP_AUX_DIR}/Calibration/calibrationlist.LowGain.dat" ]]; then
-   cp -f -v ${VERITAS_EVNDISP_AUX_DIR}/Calibration/calibrationlist.LowGain.dat ${ODIR}/Calibration/
+   if [[ ! -e ${ODIR}/Calibration/calibrationlist.LowGain.dat ]]; then
+       cp -f -v ${VERITAS_EVNDISP_AUX_DIR}/Calibration/calibrationlist.LowGain.dat ${ODIR}/Calibration/
+   fi
 else
    echo "error - low-gain calibration list not found (${VERITAS_EVNDISP_AUX_DIR}/Calibration/calibrationlist.LowGain.dat)"
    exit
 fi
 if [[ -e "${VERITAS_EVNDISP_AUX_DIR}/Calibration/LowGainPedestals.lped" ]]; then
-   cp -f -v ${VERITAS_EVNDISP_AUX_DIR}/Calibration/LowGainPedestals.lped ${ODIR}/Calibration/
+   if [[ ! -e ${ODIR}/Calibration/LowGainPedestals.lped ]]; then
+       cp -f -v ${VERITAS_EVNDISP_AUX_DIR}/Calibration/LowGainPedestals.lped ${ODIR}/Calibration/
+   fi
 else
    echo "error - low-gain calibration list not found (${VERITAS_EVNDISP_AUX_DIR}/Calibration/LowGainPedestals.lped)"
    exit
@@ -217,9 +221,11 @@ do
         -e "s|VVERSION|$EDVERSION|" \
         -e "s|DATABASETEXT|${DBTEXTDIR}|" \
         -e "s|VTS_DATA_DIR|${VERITAS_DATA_DIR}|" \
-        -e "s|VTS_2DATA_DIR|/lustre/fs23/group/veritas|" \
+        -e "s|VTS_2DATA_DIR|/lustre/fs23/group/veritas/data|" \
         -e "s|VTS_USER_DATA_DIR|${VERITAS_USER_DATA_DIR}|" \
         -e "s|USECALIBLIST|$CALIBFILE|" "$SUBSCRIPT.sh" > "$FSCRIPT.sh"
+
+#        -e "s|VTS_2DATA_DIR|/lustre/fs24/group/cta/tmp_data|" \
 
     chmod u+x "$FSCRIPT.sh"
     echo "$FSCRIPT.sh"
